@@ -58,35 +50,35 @@
     "url": "https://github.com/ai/browserslist/issues"
   },
   "dependencies": {
-    "caniuse-db": "^1.0.30000631",
-    "electron-to-chromium": "^1.2.5"
+    "caniuse-db": "^1.0.30000639",
+    "electron-to-chromium": "^1.2.7"
   },
   "description": "Share browsers list between different front-end tools, like Autoprefixer, Stylelint and babel-env-preset",
   "devDependencies": {
-    "eslint": "^3.16.1",
+    "eslint": "^3.18.0",
     "eslint-config-postcss": "^2.0.2",
     "jest": "^19.0.2",
-    "lint-staged": "^3.3.1",
+    "lint-staged": "^3.4.0",
     "pre-commit": "^1.1.3",
-    "yaspeller": "^3.0.0"
+    "yaspeller-ci": "^0.3.0"
   },
   "directories": {},
   "dist": {
-    "shasum": "af98589ce6e7ab09618d29451faacb81220bd3ba",
-    "tarball": "https://registry.npmjs.org/browserslist/-/browserslist-1.7.6.tgz"
+    "shasum": "0bd76704258be829b2398bb50e4b62d1a166b0b9",
+    "tarball": "https://registry.npmjs.org/browserslist/-/browserslist-1.7.7.tgz"
   },
   "eslintConfig": {
-    "extends": "eslint-config-postcss/es5",
     "env": {
       "jest": true
     },
+    "extends": "eslint-config-postcss/es5",
     "rules": {
       "consistent-return": "off",
       "no-process-exit": "off",
       "valid-jsdoc": "error"
     }
   },
-  "gitHead": "12ce238c326bdebe35e3abb8dcce234860398c07",
+  "gitHead": "0875962ca176f3c130c071caba6ebd69454e374d",
   "homepage": "https://github.com/ai/browserslist#readme",
   "jest": {
     "coverageThreshold": {
