@@ -64,21 +57,22 @@
     }
   ],
   "dependencies": {
-    "mime-db": ">= 1.24.0 < 2"
+    "mime-db": ">= 1.27.0 < 2"
   },
   "description": "Compressible Content-Type / mime checking",
   "devDependencies": {
-    "eslint": "3.9.1",
-    "eslint-config-standard": "6.2.1",
-    "eslint-plugin-promise": "3.3.0",
-    "eslint-plugin-standard": "2.0.1",
+    "eslint": "3.18.0",
+    "eslint-config-standard": "7.1.0",
+    "eslint-plugin-markdown": "1.0.0-beta.4",
+    "eslint-plugin-promise": "3.5.0",
+    "eslint-plugin-standard": "2.1.1",
     "istanbul": "0.4.5",
     "mocha": "~1.21.5"
   },
   "directories": {},
   "dist": {
-    "shasum": "6daab4e2b599c2770dd9e21e7a891b1c5a755425",
-    "tarball": "https://registry.npmjs.org/compressible/-/compressible-2.0.9.tgz"
+    "shasum": "feda1c7f7617912732b29bf8cf26252a20b9eecd",
+    "tarball": "https://registry.npmjs.org/compressible/-/compressible-2.0.10.tgz"
   },
   "engines": {
     "node": ">= 0.6"
