@@ -69,39 +60,39 @@
   },
   "description": "ECMAScript spec abstract operations.",
   "devDependencies": {
-    "@ljharb/eslint-config": "^7.0.0",
+    "@ljharb/eslint-config": "^10.0.0",
     "covert": "^1.1.0",
     "editorconfig-tools": "^0.1.1",
-    "eslint": "^3.3.1",
+    "eslint": "^3.14.0",
     "foreach": "^2.0.5",
     "jscs": "^3.0.7",
-    "nsp": "^2.6.1",
+    "nsp": "^2.6.2",
     "object-is": "^1.0.1",
     "replace": "^0.3.0",
-    "safe-publish-latest": "^1.0.1",
+    "safe-publish-latest": "^1.1.1",
     "semver": "^5.3.0",
-    "tape": "^4.6.0"
+    "tape": "^4.6.3"
   },
   "directories": {},
   "dist": {
-    "shasum": "bb8a2064120abcf928a086ea3d9043114285ec99",
-    "tarball": "https://registry.npmjs.org/es-abstract/-/es-abstract-1.6.1.tgz"
+    "shasum": "dfade774e01bfcd97f96180298c449c8623fb94c",
+    "tarball": "https://registry.npmjs.org/es-abstract/-/es-abstract-1.7.0.tgz"
   },
   "engines": {
     "node": ">= 0.4"
   },
-  "gitHead": "26816c72c26914c9d55cadde812e0b1dd7ca1fe4",
+  "gitHead": "78f03da0dd27ec8575fc966447635680ab8b07c2",
   "homepage": "https://github.com/ljharb/es-abstract#readme",
   "keywords": [
     "ECMAScript",
     "ES",
-    "abstract",
-    "operation",
-    "abstract operation",
-    "JavaScript",
     "ES5",
     "ES6",
-    "ES7"
+    "ES7",
+    "JavaScript",
+    "abstract",
+    "abstract operation",
+    "operation"
   ],
   "license": "MIT",
   "main": "index.js",
