@@ -65,11 +55,11 @@
   },
   "description": "Node.js final http responder",
   "devDependencies": {
-    "eslint": "3.15.0",
-    "eslint-config-standard": "6.2.1",
-    "eslint-plugin-markdown": "1.0.0-beta.3",
-    "eslint-plugin-promise": "3.3.2",
-    "eslint-plugin-standard": "2.0.1",
+    "eslint": "3.18.0",
+    "eslint-config-standard": "7.1.0",
+    "eslint-plugin-markdown": "1.0.0-beta.4",
+    "eslint-plugin-promise": "3.5.0",
+    "eslint-plugin-standard": "2.1.1",
     "istanbul": "0.4.5",
     "mocha": "2.5.3",
     "readable-stream": "2.1.2",
