@@ -1,80 +1,80 @@
 {
   "_args": [
     [
-      {
-        "raw": "is-regex@^1.0.3",
-        "scope": null,
-        "escapedName": "is-regex",
-        "name": "is-regex",
-        "rawSpec": "^1.0.3",
-        "spec": ">=1.0.3 <2.0.0",
-        "type": "range"
-      },
-      "C:\\xampp\\htdocs\\redux\\yamaha\\node_modules\\es-abstract"
+      "is-regex@^1.0.3",
+      "/home/bimbo/projects/meloriacomunicazione.local/wizard/node_modules/es-abstract"
     ]
   ],
   "_from": "is-regex@>=1.0.3 <2.0.0",
-  "_id": "is-regex@1.0.3",
+  "_id": "is-regex@1.0.4",
   "_inCache": true,
+  "_installable": true,
   "_location": "/is-regex",
-  "_nodeVersion": "1.0.4",
-  "_npmUser": {
-    "name": "ljharb",
-    "email": "ljharb@gmail.com"
+  "_nodeVersion": "7.5.0",
+  "_npmOperationalInternal": {
+    "host": "packages-12-west.internal.npmjs.com",
+    "tmp": "tmp/is-regex-1.0.4.tgz_1487407180290_0.1654177566524595"
   },
-  "_npmVersion": "2.3.0",
+  "_npmUser": {
+    "email": "ljharb@gmail.com",
+    "name": "ljharb"
+  },
+  "_npmVersion": "4.1.2",
   "_phantomChildren": {},
   "_requested": {
-    "raw": "is-regex@^1.0.3",
-    "scope": null,
-    "escapedName": "is-regex",
     "name": "is-regex",
+    "raw": "is-regex@^1.0.3",
     "rawSpec": "^1.0.3",
+    "scope": null,
     "spec": ">=1.0.3 <2.0.0",
     "type": "range"
   },
   "_requiredBy": [
     "/es-abstract"
   ],
-  "_resolved": "https://registry.npmjs.org/is-regex/-/is-regex-1.0.3.tgz",
-  "_shasum": "0d55182bddf9f2fde278220aec3a75642c908637",
+  "_resolved": "https://registry.npmjs.org/is-regex/-/is-regex-1.0.4.tgz",
+  "_shasum": "5517489b547091b0930e095654ced25ee97e9491",
   "_shrinkwrap": null,
   "_spec": "is-regex@^1.0.3",
-  "_where": "C:\\xampp\\htdocs\\redux\\yamaha\\node_modules\\es-abstract",
+  "_where": "/home/bimbo/projects/meloriacomunicazione.local/wizard/node_modules/es-abstract",
   "author": {
     "name": "Jordan Harband"
   },
   "bugs": {
     "url": "https://github.com/ljharb/is-regex/issues"
   },
-  "dependencies": {},
+  "dependencies": {
+    "has": "^1.0.1"
+  },
   "description": "Is this value a JS regex? Works cross-realm/iframe, and despite ES6 @@toStringTag",
   "devDependencies": {
-    "covert": "1.0.0",
-    "editorconfig-tools": "~0.0.1",
-    "eslint": "~0.13.0",
-    "jscs": "~1.10.0",
-    "nsp": "~1.0.0",
-    "semver": "~4.2.0",
-    "tape": "~3.4.0"
+    "@ljharb/eslint-config": "^11.0.0",
+    "covert": "^1.1.0",
+    "editorconfig-tools": "^0.1.1",
+    "eslint": "^3.15.0",
+    "jscs": "^3.0.7",
+    "nsp": "^2.6.2",
+    "replace": "^0.3.0",
+    "semver": "^5.3.0",
+    "tape": "^4.6.3"
   },
   "directories": {},
   "dist": {
-    "shasum": "0d55182bddf9f2fde278220aec3a75642c908637",
-    "tarball": "https://registry.npmjs.org/is-regex/-/is-regex-1.0.3.tgz"
+    "shasum": "5517489b547091b0930e095654ced25ee97e9491",
+    "tarball": "https://registry.npmjs.org/is-regex/-/is-regex-1.0.4.tgz"
   },
   "engines": {
     "node": ">= 0.4"
   },
-  "gitHead": "405685f6172a14d85e7a5f45e7112c9a608813f2",
+  "gitHead": "9c718fe1b3997eb70d5cbf7a6f5f233461868a00",
   "homepage": "https://github.com/ljharb/is-regex",
   "keywords": [
+    "expression",
+    "is",
     "regex",
     "regexp",
-    "is",
-    "regular expression",
     "regular",
-    "expression"
+    "regular expression"
   ],
   "license": "MIT",
   "main": "index.js",
