@@ -1,51 +1,43 @@
 {
   "_args": [
     [
-      {
-        "raw": "mime-db@~1.26.0",
-        "scope": null,
-        "escapedName": "mime-db",
-        "name": "mime-db",
-        "rawSpec": "~1.26.0",
-        "spec": ">=1.26.0 <1.27.0",
-        "type": "range"
-      },
-      "C:\\xampp\\htdocs\\redux\\yamaha\\node_modules\\mime-types"
+      "mime-db@~1.27.0",
+      "/home/bimbo/projects/meloriacomunicazione.local/wizard/node_modules/mime-types"
     ]
   ],
-  "_from": "mime-db@>=1.26.0 <1.27.0",
-  "_id": "mime-db@1.26.0",
+  "_from": "mime-db@>=1.27.0 <1.28.0",
+  "_id": "mime-db@1.27.0",
   "_inCache": true,
+  "_installable": true,
   "_location": "/mime-db",
-  "_nodeVersion": "4.6.1",
+  "_nodeVersion": "4.7.3",
   "_npmOperationalInternal": {
     "host": "packages-12-west.internal.npmjs.com",
-    "tmp": "tmp/mime-db-1.26.0.tgz_1484453096877_0.39498970191925764"
+    "tmp": "tmp/mime-db-1.27.0.tgz_1489722296902_0.15233952621929348"
   },
   "_npmUser": {
-    "name": "dougwilson",
-    "email": "doug@somethingdoug.com"
+    "email": "doug@somethingdoug.com",
+    "name": "dougwilson"
   },
-  "_npmVersion": "2.15.9",
+  "_npmVersion": "2.15.11",
   "_phantomChildren": {},
   "_requested": {
-    "raw": "mime-db@~1.26.0",
-    "scope": null,
-    "escapedName": "mime-db",
     "name": "mime-db",
-    "rawSpec": "~1.26.0",
-    "spec": ">=1.26.0 <1.27.0",
+    "raw": "mime-db@~1.27.0",
+    "rawSpec": "~1.27.0",
+    "scope": null,
+    "spec": ">=1.27.0 <1.28.0",
     "type": "range"
   },
   "_requiredBy": [
     "/compressible",
     "/mime-types"
   ],
-  "_resolved": "https://registry.npmjs.org/mime-db/-/mime-db-1.26.0.tgz",
-  "_shasum": "eaffcd0e4fc6935cf8134da246e2e6c35305adff",
+  "_resolved": "https://registry.npmjs.org/mime-db/-/mime-db-1.27.0.tgz",
+  "_shasum": "820f572296bbd20ec25ed55e5b5de869e5436eb1",
   "_shrinkwrap": null,
-  "_spec": "mime-db@~1.26.0",
-  "_where": "C:\\xampp\\htdocs\\redux\\yamaha\\node_modules\\mime-types",
+  "_spec": "mime-db@~1.27.0",
+  "_where": "/home/bimbo/projects/meloriacomunicazione.local/wizard/node_modules/mime-types",
   "bugs": {
     "url": "https://github.com/jshttp/mime-db/issues"
   },
@@ -68,14 +60,14 @@
   "dependencies": {},
   "description": "Media Type Database",
   "devDependencies": {
-    "bluebird": "3.4.7",
+    "bluebird": "3.5.0",
     "co": "4.6.0",
     "cogent": "1.0.1",
-    "csv-parse": "1.1.9",
-    "eslint": "3.13.1",
-    "eslint-config-standard": "6.2.1",
-    "eslint-plugin-promise": "3.3.0",
-    "eslint-plugin-standard": "2.0.1",
+    "csv-parse": "1.2.0",
+    "eslint": "3.17.1",
+    "eslint-config-standard": "7.0.1",
+    "eslint-plugin-promise": "3.5.0",
+    "eslint-plugin-standard": "2.1.1",
     "gnode": "0.1.2",
     "istanbul": "0.4.5",
     "mocha": "1.21.5",
@@ -84,8 +76,8 @@
   },
   "directories": {},
   "dist": {
-    "shasum": "eaffcd0e4fc6935cf8134da246e2e6c35305adff",
-    "tarball": "https://registry.npmjs.org/mime-db/-/mime-db-1.26.0.tgz"
+    "shasum": "820f572296bbd20ec25ed55e5b5de869e5436eb1",
+    "tarball": "https://registry.npmjs.org/mime-db/-/mime-db-1.27.0.tgz"
   },
   "engines": {
     "node": ">= 0.6"
@@ -97,16 +89,16 @@
     "db.json",
     "index.js"
   ],
-  "gitHead": "1d9ff30e45b07a506a20f25df3a3c7106d219e24",
+  "gitHead": "c232c21378647dfbb7762410c7b025a47f114b94",
   "homepage": "https://github.com/jshttp/mime-db#readme",
   "keywords": [
-    "mime",
-    "db",
-    "type",
-    "types",
-    "database",
     "charset",
-    "charsets"
+    "charsets",
+    "database",
+    "db",
+    "mime",
+    "type",
+    "types"
   ],
   "license": "MIT",
   "maintainers": [
@@ -135,5 +127,5 @@
     "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/",
     "update": "npm run fetch && npm run build"
   },
-  "version": "1.26.0"
+  "version": "1.27.0"
 }
