@@ -1,40 +1,32 @@
 {
   "_args": [
     [
-      {
-        "raw": "mime-types@~2.1.13",
-        "scope": null,
-        "escapedName": "mime-types",
-        "name": "mime-types",
-        "rawSpec": "~2.1.13",
-        "spec": ">=2.1.13 <2.2.0",
-        "type": "range"
-      },
-      "C:\\xampp\\htdocs\\redux\\yamaha\\node_modules\\type-is"
+      "mime-types@~2.1.15",
+      "/home/bimbo/projects/meloriacomunicazione.local/wizard/node_modules/type-is"
     ]
   ],
-  "_from": "mime-types@>=2.1.13 <2.2.0",
-  "_id": "mime-types@2.1.14",
+  "_from": "mime-types@>=2.1.15 <2.2.0",
+  "_id": "mime-types@2.1.15",
   "_inCache": true,
+  "_installable": true,
   "_location": "/mime-types",
-  "_nodeVersion": "4.6.1",
+  "_nodeVersion": "4.7.3",
   "_npmOperationalInternal": {
-    "host": "packages-18-east.internal.npmjs.com",
-    "tmp": "tmp/mime-types-2.1.14.tgz_1484458141358_0.7563668976072222"
+    "host": "packages-12-west.internal.npmjs.com",
+    "tmp": "tmp/mime-types-2.1.15.tgz_1490327753615_0.3609113476704806"
   },
   "_npmUser": {
-    "name": "dougwilson",
-    "email": "doug@somethingdoug.com"
+    "email": "doug@somethingdoug.com",
+    "name": "dougwilson"
   },
-  "_npmVersion": "2.15.9",
+  "_npmVersion": "2.15.11",
   "_phantomChildren": {},
   "_requested": {
-    "raw": "mime-types@~2.1.13",
-    "scope": null,
-    "escapedName": "mime-types",
     "name": "mime-types",
-    "rawSpec": "~2.1.13",
-    "spec": ">=2.1.13 <2.2.0",
+    "raw": "mime-types@~2.1.15",
+    "rawSpec": "~2.1.15",
+    "scope": null,
+    "spec": ">=2.1.15 <2.2.0",
     "type": "range"
   },
   "_requiredBy": [
@@ -44,11 +36,11 @@
     "/serve-index",
     "/type-is"
   ],
-  "_resolved": "https://registry.npmjs.org/mime-types/-/mime-types-2.1.14.tgz",
-  "_shasum": "f7ef7d97583fcaf3b7d282b6f8b5679dab1e94ee",
+  "_resolved": "https://registry.npmjs.org/mime-types/-/mime-types-2.1.15.tgz",
+  "_shasum": "a4ebf5064094569237b8cf70046776d09fc92aed",
   "_shrinkwrap": null,
-  "_spec": "mime-types@~2.1.13",
-  "_where": "C:\\xampp\\htdocs\\redux\\yamaha\\node_modules\\type-is",
+  "_spec": "mime-types@~2.1.15",
+  "_where": "/home/bimbo/projects/meloriacomunicazione.local/wizard/node_modules/type-is",
   "bugs": {
     "url": "https://github.com/jshttp/mime-types/issues"
   },
@@ -69,21 +61,21 @@
     }
   ],
   "dependencies": {
-    "mime-db": "~1.26.0"
+    "mime-db": "~1.27.0"
   },
   "description": "The ultimate javascript content-type utility.",
   "devDependencies": {
-    "eslint": "3.13.1",
-    "eslint-config-standard": "6.2.1",
-    "eslint-plugin-promise": "3.4.0",
-    "eslint-plugin-standard": "2.0.1",
+    "eslint": "3.18.0",
+    "eslint-config-standard": "7.1.0",
+    "eslint-plugin-promise": "3.5.0",
+    "eslint-plugin-standard": "2.1.1",
     "istanbul": "0.4.5",
     "mocha": "1.21.5"
   },
   "directories": {},
   "dist": {
-    "shasum": "f7ef7d97583fcaf3b7d282b6f8b5679dab1e94ee",
-    "tarball": "https://registry.npmjs.org/mime-types/-/mime-types-2.1.14.tgz"
+    "shasum": "a4ebf5064094569237b8cf70046776d09fc92aed",
+    "tarball": "https://registry.npmjs.org/mime-types/-/mime-types-2.1.15.tgz"
   },
   "engines": {
     "node": ">= 0.6"
@@ -93,7 +85,7 @@
     "LICENSE",
     "index.js"
   ],
-  "gitHead": "d7d15e50fe6b3a2da79d855015d25efa50e9f157",
+  "gitHead": "c44863eb0463ee16f3eb04576591cc4c4d6b214c",
   "homepage": "https://github.com/jshttp/mime-types#readme",
   "keywords": [
     "mime",
@@ -127,5 +119,5 @@
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot test/test.js",
     "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter dot test/test.js"
   },
-  "version": "2.1.14"
+  "version": "2.1.15"
 }
