@@ -1,50 +1,42 @@
 {
   "_args": [
     [
-      {
-        "raw": "psl@^1.1.7",
-        "scope": null,
-        "escapedName": "psl",
-        "name": "psl",
-        "rawSpec": "^1.1.7",
-        "spec": ">=1.1.7 <2.0.0",
-        "type": "range"
-      },
-      "C:\\xampp\\htdocs\\redux\\yamaha\\node_modules\\biskviit"
+      "psl@^1.1.7",
+      "/home/bimbo/projects/meloriacomunicazione.local/wizard/node_modules/biskviit"
     ]
   ],
   "_from": "psl@>=1.1.7 <2.0.0",
-  "_id": "psl@1.1.16",
+  "_id": "psl@1.1.18",
   "_inCache": true,
+  "_installable": true,
   "_location": "/psl",
-  "_nodeVersion": "7.5.0",
+  "_nodeVersion": "7.7.4",
   "_npmOperationalInternal": {
-    "host": "packages-12-west.internal.npmjs.com",
-    "tmp": "tmp/psl-1.1.16.tgz_1485986194809_0.6854362923186272"
+    "host": "packages-18-east.internal.npmjs.com",
+    "tmp": "tmp/psl-1.1.18.tgz_1491113861950_0.7978173801675439"
   },
   "_npmUser": {
-    "name": "lupomontero",
-    "email": "lupomontero@gmail.com"
+    "email": "lupomontero@gmail.com",
+    "name": "lupomontero"
   },
-  "_npmVersion": "4.1.2",
+  "_npmVersion": "4.4.4",
   "_phantomChildren": {},
   "_requested": {
-    "raw": "psl@^1.1.7",
-    "scope": null,
-    "escapedName": "psl",
     "name": "psl",
+    "raw": "psl@^1.1.7",
     "rawSpec": "^1.1.7",
+    "scope": null,
     "spec": ">=1.1.7 <2.0.0",
     "type": "range"
   },
   "_requiredBy": [
     "/biskviit"
   ],
-  "_resolved": "https://registry.npmjs.org/psl/-/psl-1.1.16.tgz",
-  "_shasum": "dc88ca320b87ba79f6feefdf665a8d3113b499c1",
+  "_resolved": "https://registry.npmjs.org/psl/-/psl-1.1.18.tgz",
+  "_shasum": "a8f2f5465a01e8acce4ff2d72342b05c7b507d90",
   "_shrinkwrap": null,
   "_spec": "psl@^1.1.7",
-  "_where": "C:\\xampp\\htdocs\\redux\\yamaha\\node_modules\\biskviit",
+  "_where": "/home/bimbo/projects/meloriacomunicazione.local/wizard/node_modules/biskviit",
   "author": {
     "name": "Lupo Montero"
   },
@@ -54,29 +46,29 @@
   "dependencies": {},
   "description": "Domain name parser based on the Public Suffix List",
   "devDependencies": {
-    "JSONStream": "^1.3.0",
-    "browserify": "^14.0.0",
-    "eslint": "^3.14.1",
+    "JSONStream": "^1.3.1",
+    "browserify": "^14.1.0",
+    "eslint": "^3.19.0",
     "eslint-config-hapi": "^10.0.0",
     "eslint-plugin-hapi": "^4.0.0",
     "event-stream": "^3.3.4",
-    "karma": "^1.4.1",
+    "karma": "^1.5.0",
     "karma-browserify": "^5.1.1",
     "karma-mocha": "^1.3.0",
-    "karma-mocha-reporter": "^2.2.2",
-    "karma-phantomjs-launcher": "^1.0.2",
+    "karma-mocha-reporter": "^2.2.3",
+    "karma-phantomjs-launcher": "^1.0.4",
     "mocha": "^3.2.0",
     "phantomjs-prebuilt": "^2.1.14",
-    "request": "^2.79.0",
+    "request": "^2.81.0",
     "uglifyjs": "^2.4.10",
     "watchify": "^3.9.0"
   },
   "directories": {},
   "dist": {
-    "shasum": "dc88ca320b87ba79f6feefdf665a8d3113b499c1",
-    "tarball": "https://registry.npmjs.org/psl/-/psl-1.1.16.tgz"
+    "shasum": "a8f2f5465a01e8acce4ff2d72342b05c7b507d90",
+    "tarball": "https://registry.npmjs.org/psl/-/psl-1.1.18.tgz"
   },
-  "gitHead": "914bd2c376cb979822367de4c34c43d6bb363e3b",
+  "gitHead": "1c79e5d1a6aa33d4c95dbfa56040991ba874bdc1",
   "homepage": "https://github.com/wrangr/psl#readme",
   "keywords": [
     "publicsuffix",
@@ -105,5 +97,5 @@
     "test": "mocha test && karma start ./karma.conf.js --single-run",
     "watch": "mocha test --watch"
   },
-  "version": "1.1.16"
+  "version": "1.1.18"
 }
