@@ -61,36 +54,37 @@
   ],
   "dependencies": {
     "media-typer": "0.3.0",
-    "mime-types": "~2.1.13"
+    "mime-types": "~2.1.15"
   },
   "description": "Infer the content-type of a request.",
   "devDependencies": {
-    "eslint": "2.10.2",
-    "eslint-config-standard": "5.3.1",
-    "eslint-plugin-promise": "1.1.0",
-    "eslint-plugin-standard": "1.3.2",
+    "eslint": "3.19.0",
+    "eslint-config-standard": "7.1.0",
+    "eslint-plugin-markdown": "1.0.0-beta.4",
+    "eslint-plugin-promise": "3.5.0",
+    "eslint-plugin-standard": "2.1.1",
     "istanbul": "0.4.5",
     "mocha": "1.21.5"
   },
   "directories": {},
   "dist": {
-    "shasum": "e219639c17ded1ca0789092dd54a03826b817cb2",
-    "tarball": "https://registry.npmjs.org/type-is/-/type-is-1.6.14.tgz"
+    "shasum": "cab10fb4909e441c82842eafe1ad646c81804410",
+    "tarball": "https://registry.npmjs.org/type-is/-/type-is-1.6.15.tgz"
   },
   "engines": {
     "node": ">= 0.6"
   },
   "files": [
-    "LICENSE",
     "HISTORY.md",
+    "LICENSE",
     "index.js"
   ],
-  "gitHead": "f88151e69d91c5ed42e29dea78f5566403a5a7ad",
-  "homepage": "https://github.com/jshttp/type-is",
+  "gitHead": "9e88be851cc628364ad8842433dce32437ea4e73",
+  "homepage": "https://github.com/jshttp/type-is#readme",
   "keywords": [
+    "checking",
     "content",
-    "type",
-    "checking"
+    "type"
   ],
   "license": "MIT",
   "maintainers": [
